Networking and Kubernetes - James Strong, Vallery Lancey
Kubernetes has become an essential part of the daily work for most system, network, and cluster administrators today. But to work effectively together on a production-scale Kubernetes system, they must be able to speak the same language. This book provides a clear guide to the layers of complexity and abstraction that come with running a Kubernetes network.

Authors James Strong and Vallery Lancey bring you up to speed on the intricacies that Kubernetes has to offer for large container deployments. If you're to be effective in troubleshooting and maintaining a production cluster, you need to be well versed in the abstraction provided at each layer. This practical book shows you how.

+  Learn the Kubernetes networking model.
+  Choose the best interface for your clusters from the CNCF Container Network Interface project.
+  Explore the networking and Linux primitives that power Kubernetes.
+  Quickly troubleshoot networking issues and prevent downtime.
+  Examine cloud networking and Kubernetes using the three major providers: Amazon Web Services, Google Cloud, and Microsoft Azure.
+  Learn the pros and cons of various network tools--and how to select the best ones for your stack.