Platform Engineering - Camille Fournier, Ian Nowland
Until recently, infrastructure was the backbone of organizations operating software they developed in-house. But now that cloud vendors run the computers, companies can finally bring the benefits of agile custom-centricity to their own developers. Adding product management to infrastructure organizations is now all the rage.

But how's that possible when infrastructure is still the operational layer of the company?

This practical book guides engineers, managers, product managers, and leaders through the shifts that modern platform-led organizations require. You'll learn what platform engineering is—and isn't—and what benefits and value it brings to developers and teams. You'll understand what it means to approach a platform as a product and learn some of the most common technical and managerial barriers to success.

With this book, you'll:
+  Cultivate a platform-as-product, developer-centric mindset.
+  Learn what platform engineering teams are and are not.
+  Start the process of adopting platform engineering within your organization.
+  Discover what it takes to become a product manager for a platform team.
+  Understand the challenges that emerge when you scale platforms.
+  Automate processes and self-service infrastructure to speed development and improve developer experience.
+  Build out, hire, manage, and advocate for a platform team.