Learning SQL, 3rd Edition - Alan Beaulieu
As data floods into your company, you need to put it to work right away—and SQL is the best tool for the job. With the latest edition of this introductory guide, author Alan Beaulieu helps developers get up to speed with SQL fundamentals for writing database applications, performing administrative tasks, and generating reports. You’ll find new chapters on SQL and big data, analytic functions, and working with very large databases.

Each chapter presents a self-contained lesson on a key SQL concept or technique using numerous illustrations and annotated examples. Exercises let you practice the skills you learn. Knowledge of SQL is a must for interacting with data. With Learning SQL, you’ll quickly discover how to put the power and flexibility of this language to work.


+  Move quickly through SQL basics and several advanced features.
+  Use SQL data statements to generate, manipulate, and retrieve data.
+  Create database objects, such as tables, indexes, and constraints with SQL schema statements.
+  Learn how datasets interact with queries; understand the importance of subqueries.
+  Convert and manipulate data with SQL’s built-in functions and use conditional logic in data statements.