JavaScript: The Definitive Guide, 7th Edition - David Flanagan
JavaScript is the programming language of the web and is used by more software developers today than any other programming language. For nearly 25 years this best seller has been the go-to guide for JavaScript programmers. The seventh edition is fully updated to cover the 2020 version of JavaScript, and new chapters cover classes, modules, iterators, generators, Promises, async/await, and metaprogramming. You’ll find illuminating and engaging example code throughout.

This book is for programmers who want to learn JavaScript and for web developers who want to take their understanding and mastery to the next level. It begins by explaining the JavaScript language itself, in detail, from the bottom up. It then builds on that foundation to cover the web platform and Node.js.

Topics include:
+  Types, values, variables, expressions, operators, statements, objects, and arrays.
+  Functions, classes, modules, iterators, generators, Promises, and async/await.
+  JavaScript’s standard library: data structures, regular expressions, JSON, i18n, etc.
+  The web platform: documents, components, graphics, networking, storage, and threads.
+  Node.js: buffers, files, streams, threads, child processes, web clients, and web servers.
+  Tools and language extensions that professional JavaScript developers rely on.